* serialization: '<IsSteamCloudSupported value="true" />' instead of '<IsSteamCloudSupported>true</IsSteamCloudSupported>'
