+++ Serializiation/DataTypes.cs
@@ -22,42 +22,6 @@ namespace XElement.CloudSyncHelper.Serialization.DataTypes
         public string SourceId { get; set; }
     }
 
-    public abstract class AbstractProgramInfo : IProgramInfo
-    {
-        public AbstractProgramInfo()
-        {
-            this.Configuration = new Configuration();
-        }
-
-        [XmlIgnore]
-        public Configuration Configuration { get; set; }
-        IConfiguration IProgramInfo.Configuration { get { return this.Configuration; } }
-
-        [XmlAttribute( "DisplayName" )]
-        public string DisplayName { get; set; }
-
-        [XmlAttribute( "FolderName" )]
-        public string FolderName { get; set; }
-
-        [XmlElement( "OS" )]
-        public List<OsConfiguration> OsConfigs
-        {
-            get { return this.Configuration.OsConfigs; }
-            set { this.Configuration.OsConfigs = value; }
-        }
-
-        // TODO: '<IsSteamCloudSupported value="true" />' instead of '<IsSteamCloudSupported>true</IsSteamCloudSupported>'
-        [XmlElement( "IsSteamCloudSupported" )]
-        public bool SupportsSteamCloud
-        {
-            get { return this.Configuration.SupportsSteamCloud; }
-            set { this.Configuration.SupportsSteamCloud = value; }
-        }
-
-        [XmlAttribute( "TechNameMatcher" )]
-        public string TechnicalNameMatcher { get; set; }
-    }
-
     public class AppInfo : AbstractProgramInfo, IAppInfo { }
 
     public class Configuration : IConfiguration
